pyproject.toml
[project]
name = "template-langgraph"
version = "0.0.1"
description = "A GitHub template repository for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.23.1",
    "azure-search-documents>=11.5.3",
    "elasticsearch>=9.1.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "langchain-azure-ai>=0.1.4",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.2",
    "langgraph-supervisor>=0.0.29",
    "openai>=1.98.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.15.1",
    "streamlit>=1.48.0",
    "typer>=0.16.0",
    "youtube-transcript-api>=1.2.2",
]

[dependency-groups]
docs = [
    "mkdocs-material>=9.6.14",
    "mkdocs-static-i18n>=1.3.0",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.8.0",
    "jupyterlab>=4.4.2",
    "langgraph-cli[inmem]>=0.3.6",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "ty>=0.0.1a6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.report]
show_missing = true

[tool.ty]
environment = { python-version = "3.10" }

[tool.ty.rules]
unknown-argument = "ignore"
invalid-parameter-default = "ignore"
non-subscriptable = "ignore"
possibly-unbound-attribute = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-type-form = "ignore"
invalid-assignment = "ignore"
